requirements.txt
pandas
langchain
faiss-cpu
sentence-transformers
openai
python-dotenv
streamlit
ctransformers
langchain-community
